add max audio transcription duration
Add a configurable max duration for the audio transcription, so that we have
some safeguards around the audio sent to the OpenAI API, and we don't
inadvertently incur massive API costs.

If the max duration is exceeded, simply drop the audio and do not transcribe
it. There can be more elegant ways of dealing with the audio, such as
truncating it, but this is more of a safeguard right now than a real feature.
In the future, if we understand what can cause unexpectedly long audio
durations, we can improve this accordingly.