remove cli arg parsing from pydantic settings
Parsing CLI args with Pydantic settings is not compatible with argparse, which
means that `cli.py` is broken. When running `cli.py`, we get an error saying
that the file path is not a recognized arg.

Restore `cli.py` by removing CLI arg parsing from Pydantic settings altogether.
We never really needed it anyway - we can use env vars.